requirements.txt
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.0.0
tqdm==4.64.1
segmentation-models-pytorch==0.3.3